requirements.txt
absl-py==0.11.0
addict==2.4.0
antlr4-python3-runtime==4.8
attributee==0.1.5
attrs==21.4.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
channelnorm-cuda==0.0.0
chardet==4.0.0
chumpy==0.70
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
correlation-cuda==0.0.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==5.0.3
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9c7f8a142216ebc52d3617c11f8fafd75b74e637
dotty-dict==1.3.0
fancycompleter==0.9.1
flake8==4.0.1
flake8-import-order==0.18.1
future==0.18.2
fvcore==0.1.5.post20210415
google-auth==1.27.1
google-auth-oauthlib==0.4.3
grpcio==1.36.1
hydra-core==1.1.0.dev4
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.2
iniconfig==1.1.1
iopath==0.1.8
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.1
json-tricks==3.15.5
kiwisolver==1.3.1
lap==0.4.0
loguru==0.5.3
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
mmcls @ file:///tmp/tmp16pzc8sv/mmclassification
mmcv-full==1.3.18
mmdet @ file:///tmp/tmpmu90ynq0/mmdetection
mmpose @ file:///tmp/tmp1kflnlhl/mmpose
mmtrack @ file:///tmp/tmpzy5hdacm/mmtracking
model-index==0.1.11
motmetrics==1.2.0
munkres==1.1.4
numpy==1.20.1
oauthlib==3.1.0
omegaconf==2.1.0.dev22
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
openmim==0.1.5
ordered-set==4.1.0
packaging==21.3
pandas==1.2.3
parse==1.19.0
parso==0.7.1
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==1.0.0
portalocker==2.2.1
prompt-toolkit==3.0.18
protobuf==3.15.5
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.8.0
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.8.1
pyparsing==2.4.7
pyrepl==0.9.0
pytest==7.1.1
pytest-benchmark==3.4.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resample2d-cuda==0.0.0
rich==10.1.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
setuptools-scm==6.4.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==2.1.0
tomli==2.0.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wmctrl==0.3
xmltodict==0.12.0
xtcocotools==1.11.5
yacs==0.1.8
yapf==0.30.0
zipp==3.4.1
